[IMP] coupon_criteria_multi_product, sale_coupon_criteria_multi_product
