+++ coupon_criteria_multi_product/README.rst
@@ -14,10 +14,10 @@ Coupons multi product criteria
     :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
     :alt: License: AGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fsale--promotion-lightgray.png?logo=github
-    :target: https://github.com/OCA/sale-promotion/tree/15.0/sale_coupon_criteria_multi_product
+    :target: https://github.com/OCA/sale-promotion/tree/15.0/coupon_criteria_multi_product
     :alt: OCA/sale-promotion
 .. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/sale-promotion-15-0/sale-promotion-15-0-sale_coupon_criteria_multi_product
+    :target: https://translation.odoo-community.org/projects/sale-promotion-15-0/sale-promotion-15-0-coupon_criteria_multi_product
     :alt: Translate me on Weblate
 .. |badge5| image:: https://img.shields.io/badge/runboat-Try%20me-875A7B.png
     :target: https://runboat.odoo-community.org/webui/builds.html?repo=OCA/sale-promotion&target_branch=15.0
@@ -25,8 +25,10 @@ Coupons multi product criteria
 
 |badge1| |badge2| |badge3| |badge4| |badge5| 
 
-Module that extends from *coupon_criteria_multi_product* and allows to
-use it's configuration in sale orders.
+Base module that allows to define complex rules on products to fulfill the conditions
+of a coupon program, so for a given set of products we could define a minimum quantity
+and another for others, being all those criterias mandatory for the coupon to be
+applied.
 
 **Table of contents**
 
@@ -36,7 +38,51 @@ use it's configuration in sale orders.
 Configuration
 =============
 
-Follow the steps described in the base module.
+To configure multiple product criterias:
+
+#. Go to *Sales > Catalog > Coupon Programs* and select or create a new one.
+#. On the *Coupon Criteria* field choose *Multi Product*.
+#. The standard domain will be hidden and a criteria list will be shown.
+#. Add as many criterias as needed for the desired promotion behavior.
+
+In the list of criterias we can configure:
+
+- Qty: The minimum quantity of products in the sale order to fulfill the condition.
+- Products: The products that should be present to fulfill the condition.
+- Repeat: The sum of quantities of any product in the criteria must be at least the one
+  defined in the minimum quantities. If not set, the minimum quantity will be the number
+  of products defined in the criteria, as all of them must be in the order to fulfill
+  the condition.
+
+Some examples:
+
+ ===== ================ ========
+  Qty      Products      Repeat
+ ===== ================ ========
+    1   Prod A
+ ===== ================ ========
+
+A unit of Product A must be in the sale order.
+
+ ===== ================ ========
+  Qty      Products      Repeat
+ ===== ================ ========
+    2   Prod B, Prod C
+ ===== ================ ========
+
+A unit of Product B and Product C must be in the sale order.
+
+ ===== ================ ========
+  Qty      Products      Repeat
+ ===== ================ ========
+    3   Prod D, Prod E   X
+ ===== ================ ========
+
+Either Product D or Product E or both must be in the sale order and the sum of their
+quantities must be three.
+
+Also note that all the defined criterias must be fulfilled or the program won't be
+applied.
 
 Usage
 =====
@@ -49,7 +95,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/OCA/sale-promotion/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us smashing it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/sale-promotion/issues/new?body=module:%20sale_coupon_criteria_multi_product%0Aversion:%2015.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/OCA/sale-promotion/issues/new?body=module:%20coupon_criteria_multi_product%0Aversion:%2015.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -73,7 +119,6 @@ Contributors
 * `Domatix <https://www.domatix.com>`_:
 
   * Álvaro López Oró
-  * Carolina Ferrer
 
 Maintainers
 ~~~~~~~~~~~
@@ -96,6 +141,6 @@ Current `maintainer <https://odoo-community.org/page/maintainer-role>`__:
 
 |maintainer-chienandalu| 
 
-This module is part of the `OCA/sale-promotion <https://github.com/OCA/sale-promotion/tree/15.0/sale_coupon_criteria_multi_product>`_ project on GitHub.
+This module is part of the `OCA/sale-promotion <https://github.com/OCA/sale-promotion/tree/15.0/coupon_criteria_multi_product>`_ project on GitHub.
 
 You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
