+++ coupon_criteria_multi_product/__init__.py
@@ -1,0 +1,1 @@
+from . import models
