+++ coupon_criteria_multi_product/__manifest__.py
@@ -1,7 +1,7 @@
 # Copyright 2021 Tecnativa - David Vidal
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 {
-    "name": "Coupons multi product criteria in sale",
+    "name": "Coupons multi product criteria",
     "summary": "Allows to set as promotion criteria multi-product conditions",
     "version": "15.0.1.0.0",
     "development_status": "Production/Stable",
@@ -10,5 +10,6 @@
     "author": "Tecnativa, Odoo Community Association (OCA)",
     "maintainers": ["chienandalu"],
     "license": "AGPL-3",
-    "depends": ["coupon_criteria_multi_product", "sale_coupon"],
+    "depends": ["coupon"],
+    "data": ["views/sale_coupon_program_views.xml", "security/ir.model.access.csv"],
 }
