+++ coupon_criteria_multi_product/models/__init__.py
@@ -1,0 +1,2 @@
+from . import coupon_criteria
+from . import coupon_program
