+++ coupon_criteria_multi_product/models/coupon_criteria.py
@@ -4,8 +4,8 @@ from odoo import _, api, fields, models
 from odoo.exceptions import ValidationError
 
 
-class SaleCouponCriteria(models.Model):
-    _name = "sale.coupon.criteria"
+class CouponCriteria(models.Model):
+    _name = "coupon.criteria"
     _description = "Coupon Multi Product Criteria"
 
     program_id = fields.Many2one(
@@ -21,7 +21,6 @@ class SaleCouponCriteria(models.Model):
     product_ids = fields.Many2many(
         comodel_name="product.product",
         required=True,
-        string="Products",
     )
     repeat_product = fields.Boolean(
         string="Repeat",
@@ -32,12 +31,12 @@ class SaleCouponCriteria(models.Model):
     def _compute_rule_min_quantity(self):
         """Set the minimum quantity automatically to prevent errors when the rule
         isn't set to no repeat"""
-        for criteria in self.filtered(lambda x: not x.repeat_product):
+        for criteria in self.filtered(lambda x: x.product_ids and not x.repeat_product):
             criteria.rule_min_quantity = len(criteria.product_ids)
 
     @api.constrains("rule_min_quantity")
     def _check_rule_min_qty(self):
-        for criteria in self.filtered(lambda x: not x.repeat_product):
+        for criteria in self.filtered(lambda x: x.product_ids and not x.repeat_product):
             if len(criteria.product_ids) != criteria.rule_min_quantity:
                 raise ValidationError(
                     _(
