+++ coupon_criteria_multi_product/readme/DESCRIPTION.rst
@@ -1,0 +1,4 @@
+Base module that allows to define complex rules on products to fulfill the conditions
+of a coupon program, so for a given set of products we could define a minimum quantity
+and another for others, being all those criterias mandatory for the coupon to be
+applied.
