+++ sale_coupon_criteria_multi_product/i18n/es.po
@@ -7,80 +7,82 @@ msgstr ""
 "Project-Id-Version: Odoo Server 15.0\n"
 "Report-Msgid-Bugs-To: \n"
 "POT-Creation-Date: 2023-03-03 09:06+0000\n"
-"PO-Revision-Date: 2023-03-03 09:06+0000\n"
+"PO-Revision-Date: 2023-03-03 10:08+0100\n"
 "Last-Translator: \n"
 "Language-Team: \n"
+"Language: es_ES\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
+"Content-Transfer-Encoding: 8bit\n"
+"Plural-Forms: nplurals=2; plural=(n != 1);\n"
+"X-Generator: Poedit 3.2.2\n"
 
 #. module: sale_coupon_criteria_multi_product
 #: model:ir.model.fields,help:sale_coupon_criteria_multi_product.field_coupon_program__repeat_product
 msgid "Can product quantities count multiple times or not"
-msgstr ""
+msgstr "Las cantidades de productos pueden contarse varias veces o no"
 
 #. module: sale_coupon_criteria_multi_product
 #: model:ir.model,name:sale_coupon_criteria_multi_product.model_coupon_program
 msgid "Coupon Program"
-msgstr ""
+msgstr "Programa de cupones"
 
 #. module: sale_coupon_criteria_multi_product
 #: model:ir.model.fields,field_description:sale_coupon_criteria_multi_product.field_coupon_program__create_uid
 msgid "Created by"
-msgstr ""
+msgstr "Creado por"
 
 #. module: sale_coupon_criteria_multi_product
 #: model:ir.model.fields,field_description:sale_coupon_criteria_multi_product.field_coupon_program__create_date
 msgid "Created on"
-msgstr ""
+msgstr "Creado el"
 
 #. module: sale_coupon_criteria_multi_product
 #: model:ir.model.fields,field_description:sale_coupon_criteria_multi_product.field_coupon_program__display_name
 msgid "Display Name"
-msgstr ""
+msgstr "Nombre a mostrar"
 
 #. module: sale_coupon_criteria_multi_product
 #: model:ir.model.fields,field_description:sale_coupon_criteria_multi_product.field_coupon_program__id
 msgid "ID"
-msgstr ""
+msgstr "ID"
 
 #. module: sale_coupon_criteria_multi_product
 #: model:ir.model.fields,field_description:sale_coupon_criteria_multi_product.field_coupon_program____last_update
 msgid "Last Modified on"
-msgstr ""
+msgstr "Última modificación el"
 
 #. module: sale_coupon_criteria_multi_product
 #: model:ir.model.fields,field_description:sale_coupon_criteria_multi_product.field_coupon_program__write_uid
 msgid "Last Updated by"
-msgstr ""
+msgstr "Última actualización por"
 
 #. module: sale_coupon_criteria_multi_product
 #: model:ir.model.fields,field_description:sale_coupon_criteria_multi_product.field_coupon_program__write_date
 msgid "Last Updated on"
-msgstr ""
+msgstr "Última actualización en"
 
 #. module: sale_coupon_criteria_multi_product
 #: model:ir.model.fields,field_description:sale_coupon_criteria_multi_product.field_coupon_program__rule_min_quantity
 msgid "Min. Quantity"
-msgstr ""
+msgstr "Cantidad mínima"
 
 #. module: sale_coupon_criteria_multi_product
 #: model:ir.model.fields,help:sale_coupon_criteria_multi_product.field_coupon_program__rule_min_quantity
 msgid "Minimum required product quantity to get the reward"
-msgstr ""
+msgstr "Cantidad mínima de producto requerida para obtener la recompensa"
 
 #. module: sale_coupon_criteria_multi_product
 #: model:ir.model.fields,field_description:sale_coupon_criteria_multi_product.field_coupon_program__product_ids
 msgid "Products"
-msgstr ""
+msgstr "Productos"
 
 #. module: sale_coupon_criteria_multi_product
 #: model:ir.model.fields,field_description:sale_coupon_criteria_multi_product.field_coupon_program__program_id
 msgid "Program"
-msgstr ""
+msgstr "Programa"
 
 #. module: sale_coupon_criteria_multi_product
 #: model:ir.model.fields,field_description:sale_coupon_criteria_multi_product.field_coupon_program__repeat_product
 msgid "Repeat"
-msgstr ""
+msgstr "Repetir"
