+++ sale_coupon_criteria_multi_product/models/__init__.py
@@ -1,2 +1,1 @@
-from . import sale_coupon_criteria
 from . import coupon_program
