+++ sale_coupon_criteria_multi_product/readme/DESCRIPTION.rst
@@ -1,4 +1,2 @@
-This module allows to define complex rules on products to fulfill the conditions
-of a coupon program, so for a given set of products we could define a minimum quantity
-and another for others, being all those criterias mandatory for the coupon to be
-applied.
+Module that extends from *coupon_criteria_multi_product* and allows to
+use it's configuration in sale orders.
