--- sale_coupon_criteria_multi_product/security/ir.model.access.csv
@@ -1,3 +1,0 @@
-id,name,model_id/id,group_id/id,perm_read,perm_write,perm_create,perm_unlink
-access_criteria_salesman,salesman,model_sale_coupon_criteria,sales_team.group_sale_salesman,1,0,0,0
-access_criteria_manager,criteria manager,model_sale_coupon_criteria,sales_team.group_sale_manager,1,1,1,1
