+++ setup/coupon_criteria_multi_product/odoo/addons/coupon_criteria_multi_product
@@ -1,0 +1,1 @@
+../../../../coupon_criteria_multi_product
